requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.40
accelerate>=0.27
bitsandbytes>=0.43
Pillow>=9.0
matplotlib>=3.5
numpy>=1.23
sentencepiece
protobuf